task(layout): add Portal, refactor HeaderLinks [#77]
