+++ components/common/layout/CoreLayout.js
@@ -1,6 +1,6 @@
 import './layout.scss';
 
-import React, { useState } from 'react';
+import React from 'react';
 import PropTypes from 'prop-types';
 import cn from 'classnames';
 import ScrollToTop from 'react-scroll-up';
@@ -14,7 +14,6 @@ import { LangType } from 'utils/customPropTypes';
 import Header from './Header';
 import Footer from './Footer';
 import Sidebar from './Sidebar';
-import { HeaderLinksContext, INITIAL_LINKS } from './HeaderLinks';
 
 const CoreLayout = ({ children, hideFooter, hideSidebar, lang }) => {
   const [sidebarActive, toggleSidebar, setState] = useBoolean(false);
@@ -24,21 +23,19 @@ const CoreLayout = ({ children, hideFooter, hideSidebar, lang }) => {
         id="wir-root"
         className={cn('wir-root', { 'wir-root--sidebar-expanded': sidebarActive })}
       >
-        <HeaderLinksContext.Provider value={useState(INITIAL_LINKS)}>
-          <div className="wir-space">
-            <Header toggleSidebar={toggleSidebar} />
+        <div className="wir-space">
+          <Header toggleSidebar={toggleSidebar} />
 
-            <main className="wir-content">{children}</main>
+          <main className="wir-content">{children}</main>
 
-            {!hideFooter && <Footer />}
+          {!hideFooter && <Footer />}
 
-            <Clickable
-              tag="div"
-              className={cn('wir-overlay', { 'wir-overlay--active': sidebarActive })}
-              onClick={toggleSidebar}
-            />
-          </div>
-        </HeaderLinksContext.Provider>
+          <Clickable
+            tag="div"
+            className={cn('wir-overlay', { 'wir-overlay--active': sidebarActive })}
+            onClick={toggleSidebar}
+          />
+        </div>
 
         {(!hideSidebar || sidebarActive) && (
           <nav className={cn('wir-sidebar', { 'wir-sidebar--expanded': sidebarActive })}>
