+++ components/common/layout/header/Motto.js
@@ -1,0 +1,47 @@
+import React from 'react';
+
+import Text from 'components/common/Text';
+import Link from 'components/common/Link';
+
+import { TagLink } from 'utils/tags';
+import { ROUTES_NAMES } from 'routes';
+import { TOPIC } from 'constants/misc';
+
+const Motto = () => (
+  <>
+    <span className="navbar__title--screen-wide">
+      <Text id="common.motto">
+        {(bel, asWell, world, culture, and, history) => (
+          <>
+            <TagLink topic={TOPIC.locations} tag="belarus">
+              {bel}
+            </TagLink>
+            {asWell}
+            <TagLink topic={TOPIC.locations} tag="europe">
+              {world}
+            </TagLink>
+            <TagLink topic={TOPIC.themes} tag="art">
+              {culture}
+            </TagLink>
+            {and}
+            <TagLink topic={TOPIC.themes} tag="history">
+              {history}
+            </TagLink>
+          </>
+        )}
+      </Text>
+    </span>
+    <span className="navbar__title--screen-narrow">
+      <Text id="common.motto-short">
+        {(edu, wir) => (
+          <>
+            {edu}
+            <Link route={ROUTES_NAMES.about}>{wir}</Link>
+          </>
+        )}
+      </Text>
+    </span>
+  </>
+);
+
+export default Motto;
