+++ package.json
@@ -40,7 +40,6 @@
     "@zeit/next-bundle-analyzer": "^0.1.2",
     "@zeit/next-css": "^0.2.1-canary.4",
     "@zeit/next-sass": "^0.2.1-canary.4",
-    "bulma": "^0.7.5",
     "classnames": "^2.2.6",
     "command-line-args": "^4.0.7",
     "cookie-parser": "^1.4.4",
